Move _util into __init__
